Add code into awx_manage.py
